+++ packages/adapter-static/test/apps/spa/src/routes/+error.svelte
@@ -1,13 +1,5 @@
 <script>
-	import { browser } from '$app/env';
-	import { page, session } from '$app/stores';
-
-	if (browser) {
-		$session.count += 1;
-	}
+	import { page } from '$app/stores';
 </script>
 
 <h1>{$page.status}</h1>
-<h2>count: {$session.count}</h2>
-
-<button on:click={() => ($session.count += 1)}>+1</button>