+++ packages/kit/src/runtime/client/client.js
@@ -50,13 +50,12 @@ function update_scroll_positions(index) {
 /**
  * @param {{
  *   target: Element;
- *   session: App.Session;
  *   base: string;
  *   trailing_slash: import('types').TrailingSlash;
  * }} opts
  * @returns {import('./types').Client}
  */
-export function create_client({ target, session, base, trailing_slash }) {
+export function create_client({ target, base, trailing_slash }) {
 	/** @type {Array<((href: string) => boolean)>} */
 	const invalidated = [];
 
@@ -64,7 +63,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 		url: notifiable_store({}),
 		page: notifiable_store({}),
 		navigating: writable(/** @type {import('types').Navigation | null} */ (null)),
-		session: writable(session),
 		updated: create_updated_store()
 	};
 
@@ -101,23 +99,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 
 	/** @type {import('svelte').SvelteComponent} */
 	let root;
-
-	/** @type {App.Session} */
-	let $session;
-
-	let ready = false;
-	stores.session.subscribe(async (value) => {
-		$session = value;
-
-		if (!ready) return;
-		session_id += 1;
-
-		const current_load_uses_session = current.branch.some((node) => node?.uses.session);
-		if (!current_load_uses_session) return;
-
-		update(new URL(location.href), []);
-	});
-	ready = true;
 
 	let router_enabled = true;
 
@@ -475,7 +456,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 		const uses = {
 			params: new Set(),
 			url: false,
-			session: false,
 			dependencies: new Set(),
 			parent: false
 		};
@@ -511,7 +491,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 			});
 		}
 
-		const session = $session;
 		const load_url = new LoadURL(url);
 
 		if (node.shared?.load) {
@@ -523,10 +502,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 				get url() {
 					uses.url = true;
 					return load_url;
-				},
-				get session() {
-					uses.session = true;
-					return session;
 				},
 				async fetch(resource, init) {
 					let requested;
@@ -581,6 +556,12 @@ export function create_client({ target, session, base, trailing_slash }) {
 						'@migration task: Replace `props` with `data` stuff https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292693'
 					);
 				},
+				get session() {
+					// TODO remove this for 1.0
+					throw new Error(
+						'session is no longer available. See https://github.com/sveltejs/kit/discussions/5883'
+					);
+				},
 				get stuff() {
 					throw new Error(
 						'@migration task: Remove stuff https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292693'
@@ -620,8 +601,7 @@ export function create_client({ target, session, base, trailing_slash }) {
 
 		const changed = current.url && {
 			url: id !== current.url.pathname + current.url.search,
-			params: Object.keys(params).filter((key) => current.params[key] !== params[key]),
-			session: session_id !== current.session_id
+			params: Object.keys(params).filter((key) => current.params[key] !== params[key])
 		};
 
 		// preload modules to avoid waterfall, but handle rejections
@@ -643,7 +623,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 					!previous ||
 					(changed.url && previous.uses.url) ||
 					changed.params.some((param) => previous.uses.params.has(param)) ||
-					(changed.session && previous.uses.session) ||
 					Array.from(previous.uses.dependencies).some((dep) => invalidated.some((fn) => fn(dep))) ||
 					(previous.uses.parent && nodes_changed_since_last_render.includes(true));
 				nodes_changed_since_last_render.push(changed_since_last_render);
@@ -753,7 +732,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 									uses: {
 										params: new Set(),
 										url: false,
-										session: false,
 										dependencies: new Set(),
 										parent: false
 									}
@@ -825,7 +803,6 @@ export function create_client({ target, session, base, trailing_slash }) {
 			uses: {
 				params: new Set(),
 				url: false,
-				session: false,
 				dependencies: new Set(),
 				parent: false
 			}