+++ packages/kit/test/apps/basics/src/routes/prerendering/+page.js
@@ -1,5 +1,0 @@
-export const prerender = true;
-
-export const load = ({ session }) => {
-	return session;
-};